requirements.txt
# ==================== FastAPI 核心框架 ====================
fastapi==0.121.0
uvicorn==0.38.0

# ==================== 数据验证和配置 ====================
pydantic-core==2.41.5
pydantic-settings==2.11.0
python-dotenv==1.2.1

# ==================== 数据库相关 ====================
SQLAlchemy==2.0.44
PyMySQL==1.1.2
mysql-connector-python==9.5.0

# ==================== 安全认证 ====================
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.3





